commit for classes

+++ cw1/src/base/IntElement.java
@@ -3,7 +3,7 @@ package base;
 
 public class IntElement implements IElement {
 	private String word;
-	private float value;
+	private int value;
 
 	@Override
 	public String toString() {
